+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Improvements for handling profile and profile module async updates
+
+### Fixed
+- Fixed incorrect logic when trying to set OMP theme
 
 ## [5.0.0] - 2022-03-17
 ### Added